bugfix: Trying checking if the tables exist firsts

+++ context/app.py
@@ -35,8 +35,9 @@ def create_app():
     db.init_app(app)
     Migrate(app, db)
     with app.app_context():
-        #if tables don't match migration, create them
-        db.drop_all()
+        #check if any tables exist, if so drop them
+        if db.engine.table_names():
+            db.drop_all()
         db.create_all()
 
     scheduler.init_app(app)
